requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
openslide-python>=1.3.0